requirements.txt
opencv-python
mediapipe
python-rtmidi
PyYAML
numpy